FinnhubProducer/requirements.txt
kafka-python
websocket-client
finnhub-python
avro-python3
python-dotenv
grpcio